Feat UI script titles UI page (#15466)
relates to #15224

Implement the new software page and software titles section and table.
This is with mock data for the moment. Will integrate with the new API
in future PR.

The main areas to focus on for review are:

- new routing setup for software page
- SoftwarePage.tsx
- SoftwareTitles.tsx
- services/entities/software.ts
- new software mocks
- softwareTitlesTableCongif.tsx

> NOTE: there still needs to be some work needed on the URL query params
and pagination but I plan to handle these in a future PR when I have the
new API to integrate with.

- [x] Manual QA for all new/changed functionality
