+++ frontend/__mocks__/softwareMock.ts
@@ -1,4 +1,8 @@
 import { ISoftware } from "interfaces/software";
+import {
+  ISoftwareTitle,
+  ISoftwareTitlesResponse,
+} from "services/entities/software";
 
 const DEFAULT_SOFTWARE_MOCK: ISoftware = {
   hosts_count: 1,
@@ -12,8 +16,41 @@ const DEFAULT_SOFTWARE_MOCK: ISoftware = {
   bundle_identifier: "com.app.mock",
 };
 
-const createMockSoftware = (overrides?: Partial<ISoftware>): ISoftware => {
+export const createMockSoftware = (
+  overrides?: Partial<ISoftware>
+): ISoftware => {
   return { ...DEFAULT_SOFTWARE_MOCK, ...overrides };
 };
 
-export default createMockSoftware;
+const DEFAULT_SOFTWARE_TITLE_MOCK: ISoftwareTitle = {
+  id: 1,
+  name: "mock software 1.app",
+  versions_count: 1,
+  source: "apps",
+  hosts_count: 1,
+  versions: [
+    {
+      id: 1,
+      version: "1.0.0",
+      vulnerabilities: null,
+    },
+  ],
+};
+
+export const createMockSoftwareTitle = (
+  overrides?: Partial<ISoftwareTitle>
+): ISoftwareTitle => {
+  return { ...DEFAULT_SOFTWARE_TITLE_MOCK, ...overrides };
+};
+
+const DEFAULT_SOFTWARE_TITLES_RESPONSE_MOCK: ISoftwareTitlesResponse = {
+  counts_updated_at: "2020-01-01T00:00:00.000Z",
+  count: 1,
+  software_titles: [createMockSoftwareTitle()],
+};
+
+export const createMockSoftwareTitleReponse = (
+  overrides?: Partial<ISoftwareTitlesResponse>
+): ISoftwareTitlesResponse => {
+  return { ...DEFAULT_SOFTWARE_TITLES_RESPONSE_MOCK, ...overrides };
+};
