+++ frontend/pages/SoftwarePage/SoftwareDetailsPage/_styles.scss
