+++ frontend/pages/SoftwarePage/SoftwareDetailsPage/components/Vulnerabilities/VulnTableConfig.tsx
