+++ frontend/pages/SoftwarePage/SoftwareDetailsPage/components/Vulnerabilities/Vulnerabilities.tsx
