+++ frontend/pages/SoftwarePage/SoftwareDetailsPage/components/Vulnerabilities/_styles.scss
