+++ frontend/pages/SoftwarePage/SoftwareDetailsPage/components/Vulnerabilities/index.ts
