+++ frontend/pages/SoftwarePage/SoftwareTitles/VersionCell/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./VersionCell";
