+++ frontend/pages/SoftwarePage/SoftwareTitles/_styles.scss
@@ -1,0 +1,196 @@
+.software-titles {
+  margin-top: $pad-xxlarge;
+
+  &__count {
+    display: flex;
+    gap: 12px;
+  }
+
+  &__vuln_dropdown {
+    .Select-menu-outer {
+      width: 250px;
+      max-height: 310px;
+
+      .Select-menu {
+        max-height: none;
+      }
+    }
+
+    .Select-value {
+      padding-left: $pad-medium;
+      padding-right: $pad-medium;
+    }
+
+    .dropdown__custom-value-label {
+      width: 155px; // Override 105px for longer text options
+    }
+  }
+
+  .table-container {
+    &__header {
+      flex-direction: column-reverse; // Search bar on top
+      margin-bottom: $pad-medium;
+
+      @media (min-width: $break-md) {
+        flex-direction: row;
+      }
+    }
+
+    &__header-left {
+      flex-direction: row; // Filter dropdown aligned with count
+
+      .controls {
+        .form-field--dropdown {
+          margin: 0;
+        }
+      }
+    }
+
+    &__search-input,
+    &__search {
+      width: 100%; // Search bar across entire table
+
+      .input-icon-field__input {
+        width: 100%;
+      }
+
+      @media (min-width: $break-md) {
+        width: auto;
+
+        .input-icon-field__input {
+          width: 375px;
+        }
+      }
+    }
+
+    &__data-table-block {
+      .data-table-block {
+        .data-table__table {
+          tr {
+            .software-link {
+              opacity: 0;
+              transition: opacity 250ms;
+            }
+
+            &:hover {
+              .software-link {
+                opacity: 1;
+              }
+            }
+          }
+
+          thead {
+            .name__header {
+              width: $col-md;
+            }
+
+            .version__header {
+              width: 0;
+            }
+
+            .vulnerabilities__header {
+              display: none;
+              width: 0;
+            }
+
+            .source__header {
+              display: none;
+              width: 0;
+            }
+
+            .hosts_count__header {
+              width: auto;
+              border-right: 0;
+            }
+
+            @media (min-width: $break-md) {
+              .vulnerabilities__header {
+                display: table-cell;
+              }
+            }
+
+            @media (min-width: $break-lg) {
+              .version__header {
+                width: $col-md;
+              }
+
+              .source__header {
+                display: table-cell;
+              }
+            }
+          }
+
+          tbody {
+            .name__cell {
+              max-width: $col-md;
+
+              // Tooltip does not get cut off
+              .children-wrapper {
+                overflow: initial;
+              }
+            }
+
+            .version__cell {
+              width: 0;
+            }
+
+            .source__cell {
+              display: none;
+              width: 0;
+            }
+
+            .vulnerabilities__cell {
+              display: none;
+              width: 0;
+
+              span {
+                display: inline;
+              }
+
+              .text-muted {
+                color: $ui-fleet-black-50;
+              }
+            }
+
+            .hosts_count__cell {
+              width: auto;
+
+              .hosts-cell__wrapper {
+                display: flex;
+                align-items: center;
+                justify-content: space-between;
+
+                .hosts-cell__link {
+                  display: flex;
+                  white-space: nowrap;
+                }
+              }
+            }
+
+            @media (min-width: $break-sm) {
+              .name__cell {
+                max-width: $col-lg;
+              }
+            }
+
+            @media (min-width: $break-md) {
+              .vulnerabilities__cell {
+                display: table-cell;
+              }
+            }
+
+            @media (min-width: $break-lg) {
+              .version_cell {
+                width: $col-md;
+              }
+
+              .source__cell {
+                display: table-cell;
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
