+++ frontend/pages/SoftwarePage/SoftwareVersions/SoftwareVersions.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const baseClass = "software-versions";
+
+// interface ISoftwareVersionsProps {}
+
+const SoftwareVersions = () => {
+  return <div className={baseClass}>Software Versions</div>;
+};
+
+export default SoftwareVersions;
