+++ frontend/pages/SoftwarePage/_styles.scss
@@ -1,0 +1,59 @@
+.software-page {
+  &__header-wrap {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    height: 38px;
+
+    .button-wrap {
+      display: flex;
+      justify-content: flex-end;
+      min-width: 266px;
+    }
+  }
+
+  &__manage-automations {
+    padding: $pad-small $pad-medium;
+  }
+
+  &__header {
+    display: flex;
+    align-items: center;
+
+    .form-field {
+      margin-bottom: 0;
+    }
+  }
+
+  &__text {
+    margin-right: $pad-large;
+  }
+
+  &__title {
+    font-size: $large;
+  }
+
+  &__description {
+    margin: 0;
+    margin-bottom: $pad-large;
+    max-width: 75%;
+
+    @media (min-width: $break-md) {
+      max-width: none;
+    }
+
+    h2 {
+      text-transform: uppercase;
+      color: $core-fleet-black;
+      font-weight: $regular;
+      font-size: $small;
+    }
+
+    p {
+      color: $ui-fleet-black-75;
+      margin: 0;
+      font-size: $x-small;
+      font-style: italic;
+    }
+  }
+}
