+++ frontend/pages/SoftwarePage/components/EmptySoftwareTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./EmptySoftwareTable";
