+++ frontend/pages/SoftwarePage/components/ManageAutomationsModal/ManageAutomationsModal.tsx
