+++ frontend/pages/SoftwarePage/components/ManageAutomationsModal/_styles.scss
