+++ frontend/pages/SoftwarePage/components/ManageAutomationsModal/index.ts
