+++ frontend/pages/SoftwarePage/components/PreviewTicketModal/_styles.scss
