+++ frontend/pages/SoftwarePage/components/PreviewTicketModal/index.ts
