--- frontend/pages/software/ManageSoftwarePage/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ManageSoftwarePage";
