+++ frontend/router/paths.ts
@@ -43,6 +43,10 @@ export default {
   ADMIN_ORGANIZATION_ADVANCED: `${URL_PREFIX}/settings/organization/advanced`,
   ADMIN_ORGANIZATION_FLEET_DESKTOP: `${URL_PREFIX}/settings/organization/fleet-desktop`,
 
+  // Software pages
+  SOFTWARE_TITLES: `${URL_PREFIX}/software/titles`,
+  SOFTWARE_VERSIONS: `${URL_PREFIX}/software/versions`,
+
   EDIT_PACK: (packId: number): string => {
     return `${URL_PREFIX}/packs/${packId}/edit`;
   },
@@ -107,7 +111,6 @@ export default {
   DEVICE_USER_DETAILS_POLICIES: (deviceAuthToken: string): string => {
     return `${URL_PREFIX}/device/${deviceAuthToken}/policies`;
   },
-  MANAGE_SOFTWARE: `${URL_PREFIX}/software/manage`,
   SOFTWARE_DETAILS: (id: string): string => {
     return `${URL_PREFIX}/software/${id}`;
   },
